Created utils for Dates
